Fix clipping problem
An empty image source causes the dom to render an empty image with a
white border. I created a blank image in GIMP and used it as the
placeholder for the img. This seems to be the best workaround. The
previous img src is now the background image url for the img container.