Change input management from setState to refs
Previous commits had setState updating the value of the input fields and
the className. When trying to update the className using this method,
there was a slight lag in the update because setState is asynchronous.
To solve the problem, use refs instead to bind the inputValue to custom
objects, and then update the classname using classlist.add or remove.