Attempt to change className of Panel
The React Bootstrap component is not allowing the dom to add use the
classlist add function. I can use setState since the asynchronous call
isnt that noticable for the change in css but now the css is not
complying. I just need to work on the css for when the component is
clicked and is active