Fix component to change color of button on selection
The selection of the ad format was governed by state within the
component. Now it is based on a global variable within the constructor
that persists across the lifetime of the component. THe problem with
state is that it updates asynchronously. And we need to update ui
elements immediately upon selection.