Reducing width of looks great button
Reducing the width of the bottom button that says looks great has not
been so simple. When the width is changed, the text continues to be
aligned in the same place as if the width has not been changed. Going to
attempt to modify the parent element which is the button to format into
cell table and that will hopefully allow the text to be modified.